refactor: DTO packaging
