+++ src/main/java/kitchenpos/dto/menu/MenuRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.menu;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
