+++ src/main/java/kitchenpos/dto/menugroup/MenuGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.menugroup;
 
 import kitchenpos.domain.MenuGroup;
