+++ src/main/java/kitchenpos/dto/order/OrderLineItemResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.order;
 
 import java.util.List;
 import java.util.stream.Collectors;
