+++ src/main/java/kitchenpos/dto/product/ProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.product;
 
 import java.math.BigDecimal;
 import kitchenpos.domain.Product;
