+++ src/main/java/kitchenpos/dto/tablegroup/TableGroupResponse.java
@@ -1,10 +1,11 @@
-package kitchenpos.dto;
+package kitchenpos.dto.tablegroup;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.Table;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.dto.table.TableResponse;
 
 public class TableGroupResponse {
