+++ src/test/java/kitchenpos/acceptance/AcceptanceTest.java
@@ -12,12 +12,12 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import kitchenpos.acceptance.OrderAcceptanceTest.OrderLineItemForTest;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.dto.MenuGroupResponse;
-import kitchenpos.dto.MenuResponse;
-import kitchenpos.dto.OrderResponse;
-import kitchenpos.dto.ProductResponse;
-import kitchenpos.dto.TableGroupResponse;
-import kitchenpos.dto.TableResponse;
+import kitchenpos.dto.menu.MenuResponse;
+import kitchenpos.dto.menugroup.MenuGroupResponse;
+import kitchenpos.dto.order.OrderResponse;
+import kitchenpos.dto.product.ProductResponse;
+import kitchenpos.dto.table.TableResponse;
+import kitchenpos.dto.tablegroup.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
