+++ src/test/java/kitchenpos/acceptance/MenuAcceptanceTest.java
@@ -7,9 +7,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.dto.MenuGroupResponse;
-import kitchenpos.dto.MenuResponse;
-import kitchenpos.dto.ProductResponse;
+import kitchenpos.dto.menu.MenuResponse;
+import kitchenpos.dto.menugroup.MenuGroupResponse;
+import kitchenpos.dto.product.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
