+++ src/test/java/kitchenpos/acceptance/OrderAcceptanceTest.java
@@ -7,12 +7,12 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.dto.MenuGroupResponse;
-import kitchenpos.dto.MenuResponse;
-import kitchenpos.dto.OrderLineItemResponse;
-import kitchenpos.dto.OrderResponse;
-import kitchenpos.dto.ProductResponse;
-import kitchenpos.dto.TableResponse;
+import kitchenpos.dto.menu.MenuResponse;
+import kitchenpos.dto.menugroup.MenuGroupResponse;
+import kitchenpos.dto.order.OrderLineItemResponse;
+import kitchenpos.dto.order.OrderResponse;
+import kitchenpos.dto.product.ProductResponse;
+import kitchenpos.dto.table.TableResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
