+++ src/test/java/kitchenpos/acceptance/ProductAcceptanceTest.java
@@ -3,7 +3,7 @@ package kitchenpos.acceptance;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.dto.ProductResponse;
+import kitchenpos.dto.product.ProductResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
