+++ src/test/java/kitchenpos/acceptance/TableAcceptanceTest.java
@@ -9,11 +9,11 @@ import java.util.List;
 import java.util.Map;
 import kitchenpos.acceptance.OrderAcceptanceTest.OrderLineItemForTest;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.dto.MenuGroupResponse;
-import kitchenpos.dto.MenuResponse;
-import kitchenpos.dto.OrderResponse;
-import kitchenpos.dto.ProductResponse;
-import kitchenpos.dto.TableResponse;
+import kitchenpos.dto.menu.MenuResponse;
+import kitchenpos.dto.menugroup.MenuGroupResponse;
+import kitchenpos.dto.order.OrderResponse;
+import kitchenpos.dto.product.ProductResponse;
+import kitchenpos.dto.table.TableResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
