+++ src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -9,8 +9,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import kitchenpos.dto.TableGroupResponse;
-import kitchenpos.dto.TableResponse;
+import kitchenpos.dto.table.TableResponse;
+import kitchenpos.dto.tablegroup.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
